% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.weights.r
\name{get.weights}
\alias{get.weights}
\title{Exact weights from a ps.xgb object}
\usage{
get.weights(x)
}
\arguments{
\item{x}{An object of class \link{ps.xgb}}
}
\value{
Returns a data frame.
}
\description{
Extracts weights from a ps.xgb object, output with the unique identifier for easy merging.
}
\examples{
# See vignette for examples.

}
