% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nola_south}
\alias{nola_south}
\title{An example FSRDC dataset}
\format{
A data frame

\itemize{
\item tract_id_str. Census tract identifier
\item metarea. metropolitan area
\item c00_age12. categorical variable for age
\item c00_sex. sex
\item c00_race. categorical variable for race
\item c00_nphu. number of persons in housing unit
\item hhid. household identifier
\item sim_pik. indicator of PIK assignment
\item nola_rec. indicator of record from Orleans Parish
\item id. individual identifier
}

A data frame with 18396 rows and 10 variables
}
\usage{
data(nola_south)
}
\description{
See vignette for detailed description of the data.
}
\keyword{datasets}
