% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ps.xgb.r
\name{plot.ps.xgb}
\alias{plot.ps.xgb}
\title{Plot for a ps.xbg object}
\usage{
\method{plot}{ps.xgb}(x, ...)
}
\arguments{
\item{x}{An object of class \link{ps.xgb}}

\item{...}{Arguments to be passed to other functions}
}
\value{
Returns a \link{ggplot} object.
}
\description{
\code{ps.xgb.plot} produces a figure showing the balance criteria by iteration for a
ps.xgb object.
}
\examples{
# See vignette for examples.

}
