% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nola_south}
\alias{nola_south}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{c00_age12}{price, in US dollars}
\item{c00_nphu}{weight of the diamond, in carats}
\item{c00_race}{weight of the diamond, in carats}
\item{c00_sex}{weight of the diamond, in carats}
\item{concdis}{something}
\item{hhid}{weight of the diamond, in carats}
\item{id}{weight of the diamond, in carats}
\item{imm_conc}{weight of the diamond, in carats}
\item{metarea}{weight of the diamond, in carats}
\item{nola_rec}{weight of the diamond, in carats}
\item{res_stab}{weight of the diamond, in carats}
\item{sim_pik}{weight of the diamond, in carats}
\item{tract_id_str}{weight of the diamond, in carats}
}
}
\usage{
data(nola_south)
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
